moved hbonds.hbond_autocorrel to hydrogenbonds.hbond_autocorrel
- deprecated hbonds.hbond_autocorrel (2.0.0) and stub will be
  removed in 3.0.0
- hbonds.hbond_autocorrel module raises a deprecation warning,
  HydrogenBondAutoCorrel and find_hydrogen_donors raise deprecation
  warnings
- tests (note: tests were copied and we run full tests on the deprecated
  and the moved code, mainly to make sure that they can be called as
  expected; deprecation warnings are also tested)
